mmc-utils: Pass key_mac buffer to rpmb_get_key()
This patch fixes an issue where the key_mac buffer was not passed to
rpmb_get_key() in do_rpmb_write_key(), which could result in incorrect
key handling during RPMB key programming.

This is a resend of a patch originally submitted by Simon Lindhorst,
which was previously rejected due to formatting issues.

Fixes: 05051e40351c ("mmc-utils: Refactor RPMB key handling into a separate function")
Signed-off-by: Simon Lindhorst <s.lindhorst@wut.de>
Signed-off-by: Avri Altman <avri.altman@sandisk.com>
Reviewed-by: Avri Altman <avri.altman@sandisk.com>
Link: https://lore.kernel.org/r/20250623083300.236198-1-avri.altman@sandisk.com
Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>