mmc-utils: Makefile: Make version string to show commit date
Previously, the version string included the abbreviated SHA1 of the
latest commit (e.g., "v1.0-1-g5e67f7"), which was not very informative.
This change updates the version string to includes the date of the
latest commit instead (e.g., "v1.0-1-2024-06-29"). This makes it easier
to identify the build date at a glance.

Many mmc-utils users do not build the tool themselves, so having a clear
version string with the commit date helps greatly when debugging issues
or providing support.

Signed-off-by: Avri Altman <avri.altman@sandisk.com>
Link: https://lore.kernel.org/r/20250630072420.114967-3-avri.altman@sandisk.com
Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>